That one took quite a few rounds. In the end, this netted many gains:
- Aggregated looped SQL queries into one (impact hard to gauge on remote DB)
- Replace `validateTournament` middleware by `validateID` to combine 3 SQL queries into one (impact hard to gauge on remote DB)
- Only call and fetch every members from Discord API on first usage (they are then cached and updated in real-time through Discord's gateway, saves ~250ms on subsequent calls)
- Iterate over every member in the guild only once instead of twice for every role (saves 100ms even on a recent CPU!)

In total: ~1.1s -> ~150ms

---

Applied same techniques to /staff. Response time improved from ~1.1s to ~150ms, removed caching.

And after applying these optimizations, the code for both endpoints is near identical now. I don't think we need two separate endpoints, we could just add `userRoles` to `/staff` and it'd be identical functionality-wise. Would just need to filter out users who never logged in where necessary in the front-end.